Add files via upload
Adding templates. A template file can be specified, or the default will be loaded. Then the {variables} in the template will be replaced by those fields in the yaml spec, for each method, for each path. 

Currently, these variable are hard coded, a future improvement would be to load all the variables from the template, then run the code to go and find them. This would make the template customizable, and the generator would work for all possible variables.